requirements.txt
flask==2.3.3
openai==1.30.1
line-bot-sdk==2.4.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests>=2.28.2,<2.33
yt-dlp==2024.4.9
Pillow==10.3.0
